Allow programmer to redefine all words
Check whether a word is in our main defined words list before checking
whether it might be a special word like "if" or a number. This means the
programmer can redefine those special words, even numbers. This is
consistent with the behavior of gforth.

Also, updated so we no longer dump the module if we ran into errors.