parseComment now returns CommentAST
Before, parseComment would skip the comment and return the results of
parsing whatever token came next.

This was a problem for parse functions that relied upon being able to
examine tokens BEFORE calling parseToken on them.

For example, "if" checks each token it encounters to see if it's a
"then" before calling parseToken on it (since parseToken doesn't know
what to do with "then" - it's supposed to only parse "top level"
tokens).

The result of this was a bug where a "then" immediately following a
comment would raise an error, for example.