Fixed bugs/other issues
1. The static variable "lastChar" in getNextChar was being treated as a
char. This led to portability issues because when testing for equality
with EOF (which may not cast cleanly to a "char"). It is now an int.

2. In getNextChar, I wasn't setting lastChar to EOF when we returned
EOF, though I was setting lastChar to whatever the last character was
every other time (as long as advance was true). This wasn't causing
problems, but I fixed this inconsistency by setting lastChar to EOF
where appropriate.

3. I was calling codeGenBuiltins prior to figuring out the size of a
stack item. This meant that the built-in functions were assuming the
size of a stack item was 0. This happened to work. However, the
"malloc(0)" that appeared in the assembly output has
implementation-dependent behavior. Fixed this by figuring out the stack
item size prior to code-generation built in functions.